Move common test code, add ".out" dir
